# api/candlepin-api-spec.yaml
openapi: 3.0.2
info:
  title: Candlepin
  description: Candlepin is a subscription management server written in Java. It helps with management of software subscriptions.
  version: 3.0.0-draft
servers:
  - url: /candlepin
security:
  - basicAuth: []
paths:
...
  /subscriptions:
    get:
      description: Retrieves a list of Subscriptions
      tags:
        - Subscription
      operationId: getSubscriptions
      security: []
      responses:
        200:
          description: A list of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubscriptionDTO'
        default:
          $ref: '#/components/responses/default'

    post:
      description: Activates a Subscription
      tags:
        - Subscription
      operationId: activateSubscription
      security: []
      x-java-response:
        type: javax.ws.rs.core.Response
      parameters:
        - name: consumer_uuid
          in: query
          description: The UUID of the consumer of the subscription
          required: true
          schema:
            type: string
        - name: email
          in: query
          description: The email to send a notification to for the activation
          required: true
          schema:
            type: string
        - name: email_locale
          in: query
          description: The locale of the email to send a notification to for the activation
          required: true
          schema:
            type: string
      responses:
        202:
          description: A Subscription is being activated
        400:
          description: Invalid request such as missing email/locale or when the consumer with the given uuid was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionMessage'
              example:
                displayMessage: The consumer with the given uuid was not found
                requestUuid: c4347004-8792-41fe-a4d8-fccaa0d3898a
        default:
          $ref: '#/components/responses/default'
  /subscriptions/{id}:
    delete:
      description: Removes a Subscription
      tags:
        - Subscription
      operationId: deleteSubscription
      security: []
      parameters:
        - name: id
          in: path
          description: The ID of the subscription to remove
          required: true
          schema:
            type: string
      responses:
        204:
          description: Subscription was successfully deleted
        404:
          description: Subscription could not be found using the provided ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionMessage'
              example:
                displayMessage: Subscription could not be found using the provided ID
                requestUuid: c4347004-8792-41fe-a4d8-fccaa0d3898a
        default:
          $ref: '#/components/responses/default'
components:
  responses:
    default:
      description: An unexpected exception has occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ExceptionMessage'
          example:
            displayMessage: An unexpected exception has occurred
            requestUuid: c4347004-8792-41fe-a4d8-fccaa0d3898a
  schemas:
    TimestampedEntity:
      properties:
        created:
          type: string
          format: date-time
          example: "2020-04-23 03:23:50.927000"
        updated:
          type: string
          format: date-time
          example: "2020-04-23 03:23:50.927000"
    ExceptionMessage:
      description: An exception has occurred
      properties:
        displayMessage:
          type: string
        requestUuid:
          type: string
...
    AttributeDTO:
      description: Represents the attribute as a name-value pair
      properties:
        name:
          type: string
        value:
          type: string
    BrandingDTO:
      description: Represents the branding. It indicates that a particular engineering product ID is being rebranded by the entitlement to the given name. The type is used by clients to determine what action to take with the brand name.
      allOf:
        - $ref: '#/components/schemas/TimestampedEntity'
        - type: object
          required:
            - productId
            - name
            - type
          properties:
            productId:
              type: string
              minLength: 1
              maxLength: 255
            name:
              type: string
              minLength: 1
              maxLength: 255
            type:
              type: string
              minLength: 1
              maxLength: 32
    CertificateSerialDTO:
      description: Represents a database sequence used to ensure certificates receive unique serial numbers
      allOf:
        - $ref: '#/components/schemas/TimestampedEntity'
        - type: object
          properties:
            id:
              type: string
            serial:
              type: string
            expiration:
              type: string
              format: date-time
            collected:
              type: boolean
            revoked:
              type: boolean
    CertificateDTO:
      description: Represents certificate details
      allOf:
        - $ref: '#/components/schemas/TimestampedEntity'
        - type: object
          properties:
            id:
              type: string
            key:
              type: string
            cert:
              type: string
            serial:
              $ref: '#/components/schemas/CertificateSerialDTO'
...
    ContentDTO:
      description: Represents the content data exposed to the API.
      required:
        - id
        - uuid
      allOf:
        - $ref: '#/components/schemas/TimestampedEntity'
        - type: object
          properties:
            uuid:
              type: string
              minLength: 1
              example: "ff808081554a3e4101554a3e9033005d"
            id:
              type: string
              minLength: 1
              example: "5001"
            type:
              type: string
              example: "yum"
            label:
              type: string
              example: "content_label"
            name:
              type: string
              example: "content_name"
            vendor:
              type: string
              example: "example-vendor"
            contentUrl:
              type: string
              example: "/admin/foo/example/path"
            requiredTags:
              type: string
              example: "TAG1,TAG2"
            releaseVer:
              type: string
              example: "1.2.3"
            gpgUrl:
              type: string
              example: "/admin/foo/example/gpg/path"
            modifiedProductIds:
              type: array
              uniqueItems: true
              example: "[5051,5052,5053]"
              items:
                type: string
            arches:
              type: string
              example: "x86_64"
            metadataExpire:
              type: integer
              format: int64
              example: 1467124079
...
    CdnDTO:
      description: DTO representing a Cdn
      allOf:
        - $ref: '#/components/schemas/TimestampedEntity'
        - type: object
          properties:
            id:
              type: string
            label:
              type: string
            name:
              type: string
            url:
              type: string
            certificate:
              $ref: '#/components/schemas/CertificateDTO'
...
    NestedOwnerDTO:
      properties:
        id:
          type: string
        key:
          type: string
        displayName:
          type: string
        href:
          type: string
...
    ProductContentDTO:
      description: Product content mapping exposed to the API
      required:
        - content
      properties:
        content:
          $ref: "#/components/schemas/ContentDTO"
        enabled:
          type: boolean
    ProductDTO:
      description: DTO representing the product data exposed to the API
      allOf:
        - $ref: '#/components/schemas/TimestampedEntity'
        - type: object
          properties:
            id:
              type: string
              example: 5051
            uuid:
              type: string
              example: ff808081554a3e4101554a3e9033005d
            name:
              type: string
              example: Admin OS Developer Bits
            multiplier:
              type: integer
              format: int64
              example: 1
            attributes:
              type: array
              items:
                $ref: '#/components/schemas/AttributeDTO'
            productContent:
              type: array
              uniqueItems: true
              items:
                $ref: "#/components/schemas/ProductContentDTO"
            dependentProductIds:
              type: array
              uniqueItems: true
              items:
                type: string
            branding:
              type: array
              uniqueItems: true
              items:
                $ref: "#/components/schemas/BrandingDTO"
            href:
              type: string
              example: /products/ff808081554a3e4101554a3e9033005d
...
    SubscriptionDTO:
      description: Represents the Subscription data exposed to the API
      allOf:
        - $ref: '#/components/schemas/TimestampedEntity'
        - type: object
          properties:
            id:
              type: string
            owner:
              $ref: "#/components/schemas/NestedOwnerDTO"
            product:
              $ref: "#/components/schemas/ProductDTO"
            derivedProduct:
              $ref: "#/components/schemas/ProductDTO"
            providedProducts:
              type: array
              uniqueItems: true
              items:
                "$ref": "#/components/schemas/ProductDTO"
            derivedProvidedProducts:
              type: array
              uniqueItems: true
              items:
                $ref: "#/components/schemas/ProductDTO"
            quantity:
              type: integer
              format: int64
            startDate:
              type: string
              format: date-time
              example: "2019-05-08 09:42:37.000"
            endDate:
              type: string
              format: date-time
              example: "2019-05-08 09:42:37.000"
            contractNumber:
              type: string
            accountNumber:
              type: string
            modified:
              type: string
              format: date-time
              example: "2019-05-08 09:42:37.000"
            lastModified:
              type: string
              format: date-time
              example: "2019-05-08 09:42:37.000"
            orderNumber:
              type: string
            upstreamPoolId:
              type: string
            upstreamEntitlementId:
              type: string
            upstreamConsumerId:
              type: string
            cert:
              $ref: "#/components/schemas/CertificateDTO"
            cdn:
              $ref: "#/components/schemas/CdnDTO"
            stacked:
              type: boolean
            stackId:
              type: string
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic